# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description:
  version: 1.0.0
  title: 接口文档
basePath: /
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /bj/api/login:
    post:
      tags:
        - bj
      summary: bj登录
      produces:
        - application/json
      parameters:
        - name: username
          description: 用户名
          in: path
          type: string
          required: true
        - name: pwd
          description: 密码
          in: path
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /bj/api/xrxgjl:
    post:
      tags:
        - bj
      summary: 写入巡更记录
      produces:
        - application/json
      parameters:
        - name: body
          description: 写入巡更记录
          in: body
          type: string
          schema:
            $ref: '#/definitions/XgjlEntity'
          required: true
        - name: token
          description: token
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /bj/api/xrgps:
      post:
        tags:
          - bj
        summary: 写入GPS
        produces:
          - application/json
        parameters:
          - name: body
            description: 写入GPS
            in: body
            type: string
            schema:
              $ref: '#/definitions/GpsEntity'
            required: true
          - name: token
            description: token
            in: query
            type: string
            required: true
        responses:
          '200':
            schema:
              $ref: '#/definitions/R'

#定义数据模型
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  XgjlEntity:
    type: object
    properties:
      name:
        description: 巡更点名称
        type: string
      lng:
        description: lng
        type: string
      lat:
        description: lat
        type: string
      xgTime:
        description: 巡更时间
        type: string
      xgr:
        description: 巡更人
        type: string
  GpsEntity:
      type: object
      properties:
        name:
          description: 巡更点名称
          type: string
        lng:
          description: lng
          type: string
        lat:
          description: lat
          type: string
        xgTime:
          description: 巡更时间
          type: string
        xgr:
          description: 巡更人
          type: string